Float32 fix

+++ main.py
@@ -67,7 +67,7 @@ def parameter_search(dir='models/search/'):
 		config = json.load(data_file)
 	train_set, validation_set, test_set = dataset.load_dataset(config['no_dwis'], split_ratio=(0.6, 0.2, 0.2))
 
-	learning_rates = 10 ** np.random.uniform(-6, -3, 20)
+	learning_rates = 10 ** np.random.uniform(-6, -3, 20).astype(dtype=np.float32)
 	batch_norms = [False, True]
 	with_stds = [False, True]
 	layers = [
